Fold the shared layers into xdeepfm.py and drop the string dispatch tables
Follows the sasrec/model.py precedent: one self-contained model file per
model, with its own sub-layers, instead of a neutral layers.py shared across
model families. sequential_base.py and sli_rec.py go back to their staging
state, so this PR no longer touches the SLi-Rec sources at all.

The ACTIVATIONS dict and the init_method dispatch were leftovers of the TF
yaml/hparams design this port removed. The DNN now calls F.relu directly
(activation: [relu, relu] in every shipped config), the CIN applies no
activation (cross_activation: identity everywhere), and weights use
nn.init.trunc_normal_ directly (init_method: tnormal everywhere). The
activation, cross_activation and init_method arguments are gone.

Verified bit-identical logits against the previous commit for all four
components with batch normalization enabled.
