+++ docs/models.rst
@@ -77,9 +77,6 @@ xDeepFM
 .. automodule:: recommenders.models.deeprec.models.pytorch.xdeepfm
     :members:
     :special-members: __init__
-.. automodule:: recommenders.models.deeprec.models.pytorch.layers
-    :members:
-    :special-members: __init__
 .. automodule:: recommenders.models.deeprec.io.ffm_dataset
     :members:
     :special-members: __init__
